[CPUFREQ] speedstep-centrino: transient MSR values (Venkatesh Pallipadi)
On some CPUs, we can see transient MSR values (which are not present in _PSS)
in IA32_PERF_STATUS MSR, while CPU is doing some automatic P-state transition
(like TM2).
Current code will return frequency as 0 in such cases. Fix it by retrying the
get after a delay and use lowest possible frequency as the current frequency
in worst case.

Thanks to Matt Domsch for identifying and root-causing this failure.

Signed-off-by: Venkatesh Pallipadi <venkatesh.pallipadi@intel.com>
Signed-off-by: Dominik Brodowski <linux@brodo.de>
Signed-off-by: Dave Jones <davej@redhat.com>