[ARM PATCH] 2324/1: Convert ARM proc files to use domain and pgtable manifest constants
Patch from George G. Davis

Convert ARM proc files to use manifest constants defined in pgtable.h
and domain.h. This is a prelude to allow renumbering domains. The
domain renumbering is required to implement support for ARMv6 MMU
super sections where domain 0 is implied. Although this is technically
only required for ARMv6, it seemed reasonable to clean up all proc
files while here.

I did a test build with and without this patch applied and confirmed
that all proc file binaries are unchanged when this patch is applied.

Signed-off-by: George G. Davis
Signed-off-by: Russell King