[ARM] Add DMA mmap() support.
This adds DMA mmap() support for the ARM architecture, as discussed
around March 2003 on the linux-arch and linux-kernel mailing lists.

Subsystems such as ALSA (for sample ring buffers) and video drivers
(for framebuffers in system memory) require this infrastructure to
provide userspace with an architecture clean method to mmap these
memory areas.

Signed-off-by: Russell King <rmk@arm.linux.org.uk>