[PATCH] x86_64: Scheduler support for AMD CMP
Scheduler support for AMD CMP

AMD dual core looks like HT but isn't really.  Hide it from the scheduler. 
This works around problems with the domain scheduler.  Also probably gives
slightly better scheduling and disables SMT nice which is harmful on dual
core.

We still display it like HT in /proc/cpuinfo.  This is to make it appear the
same to license managers so that they count dual cores only as a single
physical CPU.

Signed-off-by: Andi Kleen <ak@suse.de>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>