[PATCH] uml: close host file descriptors properly
This process closes some file descriptors which were left open incorrectly.
These are the initrd descriptor, the temporary test file used for testing /tmp
for execution permission, and a descriptor used by the netork to connect to
the switch.  In the network case, we add network devices to the opened list as
soon as they are added to UML, rather than when they are configured.  This
ensures that close_devices will remove the device properly on shutdown.

Signed-off-by: Jeff Dike <jdike@addtoit.com>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>