[PATCH] ext3: cleanup handling of aborted transactions.
This patch improves ext3's error logging when we encounter an on-disk
corruption.  Previously, a transaction (such as a truncate) which encountered
many corruptions (eg.  a single highly-corrupt indirect block) would emit
copious "aborting transaction" errors to the log.

Even worse, encountering an aborted journal can count as such an error,
leading to a flood of spurious "aborting transaction: Journal has aborted"
errors.

With the fix, only emit that message on the first error.  The patch also
restores a missing \n in that printk path.

Signed-off-by: Stephen Tweedie <sct@redhat.com>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>