[PATCH] fix cdrom autoclose
The attached patch fixes the CD-ROM autoclose.  It is broken in recent
kernels for CD-ROMs that do not properly report that the tray is opened.
Now on such a drives the kernel will do one close attempt and check for the
disc again.  This is how it used to work in the past.

Signed-off-by: Stas Sergeev <stsp@aknet.ru>
Acked-by: Alexander Kern <alex.kern@gmx.de>
Acked-by: Jens Axboe <axboe@suse.de>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>