[PATCH] 3c59x: reload EEPROM values at rmmod for needy cards
Enable reload of EEPROM values in reset at rmmod for cards that need it,
similar to old EEPROM_NORESET flag but in reverse.

(Most?) 3c905 and (some?) 3c905B cards need an additional bit unmasked in
the reset at rmmod or else they don't get reinitialized properly when the
driver is reloaded.

Signed-of-by: John W. Linville <linville@tuxdriver.com>

Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>